implement GraknConsoleIT in CircleCI and update for split client and server
