+++ test/GraknConsoleIT.java
@@ -25,17 +25,14 @@ import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Sets;
 import grakn.console.GraknConsole;
-import grakn.console.exception.GraknConsoleException;
 import grakn.core.rule.GraknTestServer;
 import graql.lang.Graql;
-import io.grpc.Status;
 import org.apache.commons.io.output.TeeOutputStream;
 import org.hamcrest.Matcher;
 import org.junit.AfterClass;
 import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.ClassRule;
-import org.junit.Ignore;
 import org.junit.Test;
 
 import java.io.ByteArrayInputStream;
@@ -45,6 +42,7 @@ import java.io.InputStream;
 import java.io.OutputStream;
 import java.io.PrintStream;
 import java.nio.file.Files;
+import java.nio.file.Paths;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Stream;
@@ -67,7 +65,10 @@ import static org.junit.Assert.assertTrue;
 public class GraknConsoleIT {
 
     @ClassRule
-    public static final GraknTestServer server = new GraknTestServer();
+    public static final GraknTestServer server = new GraknTestServer(
+            Paths.get("test/resources/grakn.properties"),
+            Paths.get("test/resources/cassandra-embedded.yaml")
+    );
 
     private static InputStream trueIn;
     private static int keyspaceSuffix = 0;
@@ -145,14 +146,14 @@ public class GraknConsoleIT {
 
     @Test
     public void when_startingConsoleWithOptionLoadFile_expect_noError() {
-        Response response = runConsoleSession("", "-f", "console/test/file-(with-parentheses).gql");
+        Response response = runConsoleSession("", "-f", "test/file-(with-parentheses).gql");
         assertEquals("", response.err());
     }
 
     @Test
     public void when_loadingFileInConsoleSession_expect_dataIsWritten() throws Exception {
         assertConsoleSessionMatches(
-                "load console/test/file-(with-parentheses).gql",
+                "load test/file-(with-parentheses).gql",
                 anything(),
                 "match movie sub entity; get; count;",
                 containsString("1")
@@ -161,7 +162,7 @@ public class GraknConsoleIT {
 
     @Test
     public void when_loadingInvalidDataFromFile_expectError() {
-        Response response = runConsoleSession("", "-f", "console/test/invalid-data.cql");
+        Response response = runConsoleSession("", "-f", "test/invalid-data.cql");
 
         assertThat(response.err(), allOf(containsString("Failed to load file:"), containsString("A structural validation error has occurred.")));
         assertThat(response.out(), not(containsString("Successful commit:")));
@@ -170,7 +171,7 @@ public class GraknConsoleIT {
     @Test
     public void when_loadingFileWithEscapes_expect_dataIsWritten() throws Exception {
         assertConsoleSessionMatches(
-                "load console/test/file-\\(with-parentheses\\).gql",
+                "load test/file-\\(with-parentheses\\).gql",
                 anything(),
                 "match movie sub entity; get; count;",
                 containsString("1")
