+++ test/resources/grakn.properties
@@ -1,0 +1,79 @@
+#
+# GRAKN.AI - THE KNOWLEDGE GRAPH
+# Copyright (C) 2019 Grakn Labs Ltd
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as
+# published by the Free Software Foundation, either version 3 of the
+# License, or (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Affero General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with this program.  If not, see <https://www.gnu.org/licenses/>.
+#
+
+############################# Knowledge Base Configuration #############################
+
+# Number of instances after which Grakn should shard any type node. Sharding is the
+# mechanism by which Grakn mitigates against supernodes.
+# A larger sharding threshold will increase runtime as the knowledge base  grows and decrease
+# the likelihood of supernodes. A threshold that is too small will create supernodes
+# more frequently.
+knowledge-base.type-shard-threshold=250000
+
+############################# Server Configuration #############################
+
+# Directory in which server data will be stored
+data-dir=server/db/
+
+# Set the IP address that Grakn server server will listen on.
+server.host=0.0.0.0
+
+# Port number to use for gRPC server to listen on
+grpc.port=48555
+
+############################# Logging Configuration #############################
+# These properties are read directly by logback.xml
+
+# Single directory under which to store log files. If using a relative path it will be relative
+# to the JVM system variable "grakn.dir"
+log.dirs=./logs/
+
+# Specify the log verbosity level. This can be one of:
+#
+# ERROR: (critical errors indicating the application has failed)
+# WARN:  (errors that will not affect the overall running of the application)
+# INFO:  (minimally verbose, including Grakn server lifecycle events)
+# DEBUG: (verbose, non production use, server task lifecycle events)
+# TRACE: (extraordinarily verbose, be brave, including graql query traversal paths,
+#         extra task lifecycle events, kafka consumer offsets, etc)
+log.level=INFO
+
+############################# Persisted Knowledge Base Configuration #############################
+# The hostname or comma-separated list of hostnames of storage backend servers
+storage.hostname=127.0.0.1
+storage.port=9160
+
+#Port used by KeyspaceManager
+cassandra.input.native.port=9042
+
+# Timeout in milliseconds when connecting to storage backend
+storage.connection-timeout=20000
+
+# Frame size for thrift (maximum message length).
+storage.internal.thrift_framed_transport_size_in_mb=60
+
+# Whether to enable the database-level cache, which is shared across all transactions.
+# Enabling this option speeds up traversals by holding hot elements in memory,
+# but also increases the likelihood of reading stale data. Disabling it forces each transaction
+# to independently fetch elements from storage before reading/writing them.
+# the default cache expiration is 5 seconds, currently unmodifiable
+cache.db-cache=false
+# Size of Janus's database cache in proportion to JVM size 0 (small) to 1 (large)
+cache.db-cache-size=0.35
+cache.tx-cache-size=30000
+cache.tx-dirty-size=4096
